Minor optimisation
